Fix Package-Requires line
The existing line was not package.el-compliant, and rendered the
file unparseable by the `package-buffer-info` function. This
commit fixes the line, and additionally modifies the Carton file
to use the new "package-file" directive: using this directive,
Carton will read the package dependencies and version number from
the .el file just like package.el, meaning that changes need only
be made in one place.